refactor: unify services context
